refactor: organize parser error tests by associated modules
Move parser error tests from error.rs to their most closely associated
modules for better discoverability and organization:
- items.rs: vftable, type definition, enum, bitflags tests
- functions.rs: impl block, function signature tests
- external.rs: use statement, extern type tests
- attributes.rs: attribute parsing/bounds checking tests
- module.rs: module-level token error tests
- types.rs: pointer, array, generic type error tests
